+++ trunk/sfox/src/starengine/engine/frustum.h
@@ -29,7 +29,7 @@ typedef struct {
 } frustum2d;
 
 extern void frustum_get_from_opengl(frustum *ftm);
-extern int frustum_point_is_into(frustum *ftm, double x, double y, double z);
+extern int frustum_point_is_into(frustum *ftm, float x, float y, float z);
 extern void frustum_copy(frustum *dst, frustum *src);
 /*Only ftm is modified*/
 extern void frustum_transform(frustum *ftm, matrix4 m, vector3 *pos);
@@ -39,7 +39,7 @@ extern void frustum_print(frustum *ftm);
 
 /*Project the frustum on the xy plane*/
 extern void frustum_to_frustum2d_xz(frustum2d *ftm2d, frustum *ftm);
-extern int frustum2d_point_is_into(frustum2d *ftm, double x, double y);
+extern int frustum2d_point_is_into(frustum2d *ftm, float x, float y);
 extern int frustum2d_bbox_is_into(frustum2d *ftm2d, bbox2d *bb);
 
 #endif
