+++ trunk/sfox/src/starengine/engine/material.h
@@ -9,7 +9,7 @@ enum { DIFFUSE=1, SPECULAR=2, TEXTURE=4, COLOR=8 };
 typedef struct material *material;
 
 /* col is copied into material, tex[] contains no more than MAX_TEXTURES tetures */
-extern material material_create(texture tex[], int num_textures, color *col, double diffuse, double specular);
+extern material material_create(texture tex[], int num_textures, color *col, float diffuse, float specular);
 
 /*Create a material which is only a texture*/
 extern material material_create_single_texture_from_file(char *filename);
