+++ trunk/sfox/src/starengine/engine/vertex.h
@@ -15,10 +15,10 @@ typedef struct vertex {
 #define VERTEXOFFSETTCOORD(i) (sizeof(vector3)+sizeof(vector2)*(i))
 #define VERTEXOFFSETCOLOR (sizeof(vector3)+sizeof(vector2)*(MAX_TEXTURES))
 
-extern void vertex_set_coord(vertex *v, double x, double y, double z);
-extern void vertex_set_color(vertex *v, double r, double g, double b, double a);
+extern void vertex_set_coord(vertex *v, float x, float y, float z);
+extern void vertex_set_color(vertex *v, float r, float g, float b, float a);
 /* Set texture coordinates for texture n*/
-extern void vertex_set_tcoord(vertex *vx, int n, double u, double v);
+extern void vertex_set_tcoord(vertex *vx, int n, float u, float v);
 extern void vertex_copy(vertex *dest, vertex *src);
